+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- move `save-dependencies-as-components` flag from `bit import` command to be configurable in consumer bit.json
+
 ## [0.12.0-ext.10] - 2017-12-27
 
 - add unsafe-perm for installing packages inside components